requirements.txt
fastapi==0.115.6
python-socketio==5.12.1
asyncpg==0.30.0
redis==5.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
httpx==0.28.1
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
